+++ micc/template-package(obsolete)/{{cookiecutter.project_name}}/tests/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Unit test package for {{ cookiecutter.project_name }}.
-"""